requirements.txt
numpy
dotmap
rich
google-cloud-storage
paramiko
sshtunnel
